[XLA:GPU] Allow fusing multiply into triton fusion.
The pr enables multiply fusion for the cases when its arguments have positive effect on the input sizes. I.e. when total size of multiply args is less than the output size of the multiply. It is the case when we multiply s4 by a small number of scales and pass the result to the dot.

PiperOrigin-RevId: 716256953