[XLA:MSA] Fix the AllocationRequest for window prefetch
We currently have two performance issues in window prefetch.

The first is that we specified the created WindowPrefetchedAllocation to consume CopyResources by using non-zero shape. This could consume all resources and interfere with the prefetching decision of other tensors. In fact, we don't have prefetching implemented yet, so we could specify using zero CopyResource.

Another issue is that the generated allocation from window prefetch currently spans too long in time. Its earliest prefetch time is set to the operand's define time, and its end is the use time. We should keep the earliest prefetch time to be as close as the use time as possible. This is for keeping the interference to the prefetching of other tensors to be minimal.

We updated WindowPrefetch() to simply just allocate chunk for the exposed span vmem at this moment. Since we don't call Prefetch() from WindowPrefetch(), we can simplify the data structure of AllocationRequest and PrefetchContext a bit.

PiperOrigin-RevId: 718439738