[xla:gpu] [cleanup] Enclose iterable inputs into their own class
This brings down the "cognitive complexity" of the function from 78 to 60.

I am working on adding TMA support to EmitMatmul, but this function is cannot support it the way it is currently written. EmitMatmul assumes that each input can be easily translated to an iterable argument, which is not the case for TMA where we are iterating over 2 indices rather than a tensor pointer.

Previously, the function relied on multiple iter_arg_to-* functions. This change refactors the code to enclose each input that needs to be emitted on each iteration of the ForOp into its own class. This can further be abstracted to fully emit the parameter value & calculate indices next.

Previously, there were many different index calculations happening here & I tried to abstract them to be named in order for it to be more legible.

PiperOrigin-RevId: 720151542