[XLA:CPU] Add multithread support to FFT thunk.
As reported in https://github.com/jax-ml/jax/issues/25808, the performance of XLA's CPU FFT is dramatically reduced with the thunks runtime. This is because the intra-op thread pool wasn't properly passed through. This change adds multithreading support, by passing through the thread pool provided by xla::ExecuteParams, but I don't know if this is the right approach.

PiperOrigin-RevId: 720166329